add gripper_center_link to rexarm URDF
gripper_center_link corresponds to the center of the gripper, oriented
in the same way as the gripper is facing. so a neutral orientation with
all joints at zero corresponds to gripper_center_link having x facing
straight up. in other words, the x-axis of gripper_center_link is the
direction that the gripper is looking at or facing